refactor: split App core into hooks + ChallengePanel (move-only)
Final step of the App.tsx god-component split. App is now a thin shell:
heading, the five-tab view switch, and the panel router. Everything else
moves out behind a small interface:

- src/hooks/useTheme.ts          - theme state, the <html data-theme>
                                   effect, and the persisted toggle.
- src/hooks/usePracticeSession.ts - the stateful practice core: all
                                   session state, the derived question
                                   pool, and every drill handler. Returns
                                   one session object.
- src/components/ChallengePanel.tsx - the three-column practice layout
                                   (controls / drill / mistake review).
                                   Takes the session object + language +
                                   an onExit callback; no local state.

App keeps appView and composes the "start drill / go to challenge"
actions from the session setters, so the hooks stay free of view routing.

No behavioural change - pure move + rewiring. The full App integration
suite (challenge flow, drills, review finite-pass, mock picker, theme)
passes unchanged.

App.tsx: 803 -> 172 lines (1542 before the whole split).

Verification: tsc, vitest (118/118), vite build, check:exam all green.
