+++ src/components/index.ts
@@ -1,5 +1,6 @@
 // Barrel for the view panels + leaf presentational components extracted
 // from App.tsx. Keeps App's import list short; add future components here.
+export { ChallengePanel } from "./ChallengePanel";
 export { HomePanel } from "./HomePanel";
 export { LearningPanel } from "./LearningPanel";
 export { RulesPanel } from "./RulesPanel";
